Add more tree mutations and expose commands
This commit adds the following commands:
* AerialTreeOpen
* AerialTreeClose
* AerialTreeToggle
* AerialTreeOpenAll (new functionality)
* AerialTreeCloseAll (new functionality)
* AerialTreeSyncFolds

With the new functionality, we are now binding `zM` and `zR` in the aerial buffer, as well as in normal buffers when `link_folds_to_tree = true`.

Additionally I extracted the tree mutation logic into tree.lua and simplified some of the folding logic.