Bug fix: closing wrong window when switching buffers
This was happening as we were creating a new aerial buffer for the open window.
We enter the buffer before the filetype is set (we have to, otherwise the
FileType autocmds won't trigger properly), and doing so would cause this autocmd
to close the window and replace the main window with the new empty aerial
buffer.