Fix empty aerial buffer when using other plugins
Some other plugins make use of floating windows (e.g. telescope,
nvim-treesitter-context), and that was causing some issues when
close_behavior = 'global'. When the floating window is opened, the
aerial window thinks that it's no longer relevant and clears itself.
This can range from annoying to UI breaking, depending on how the
floating window is used.