fix(treesitter): disable parsing of symbols from subtrees
The issue is that the injected trees are all processed *after* the top
level tree, so the symbols are just tacked on to the end of the list
regardless of the subtree's position in the document. This produces an
incorrect outline and also messes up the navigation commands. More work
will need to be done to figure out how to place subtrees correctly.