perf: improve performance for large files
This field was unused, so we can remove it. The trick here is that
computing it calls out to `config`, which does a truly unfortunate
amount of work. For single calls, it's fine. For files with >1000
symbols, it gets slow. I'm likely going to have to change how aerial is
configured because the global var lookup will always have this issue.