refactor: Change aerial config to use setup function (#37)
This change is backwards-compatible, though the old global-variable
method will be removed at some point. The motivation for this change is
that looking up (and potentially processing) the values from global
variable on each call to `config.<var>` is slow (in a way that matters
for large files). We're fixing this by requiring a call to `setup()`
before using aerial and preprocessing all the options that need it. Now
config lookups are as fast as a table access.