feat: add symbol ranges for treesitter and markdown (#52)
This adds support for the end_lnum and end_col fields to the treesitter
and markdown backends. Now all backends store both the start and the end
of a symbol. At the moment this is only being used for detecting exact
position matching, so the lualine component and highlighting when
`highlight_closest = false`.

This change also made the C, C++, and Julia treesitter backends a bit
more correct and robust.