refactor: prefer treesitter backend over LSP
The Treesitter backend has been around long enough that I'm convinced
it's stable. Furthermore, I think that in general the usability is
better (it's faster and supports more customization e.g. see the
difference in rust symbols). This is easy for users to change in their
own config, but ideally if the LSP backend is providing something the
Treesitter backend doesn't, TS should be updated to reach feature
parity.