refactor!: Link aerial to a source window, not a buffer (#128)
This is a follow-up to #144, which did a lot of work to make sure that
aerial could logically be opened for a *window* and would stay in sync
with whatever was displayed in that window. Some of the aerial logic was
still checking for linked aerial *buffers* instead of a linked *window*,
which led to some non-intuitive behavior. This change should mostly make
things behave in a less surprising way, but it *does* change the
behavior in some cases so it *might* cause perceived bugs.

Since we've moved more towards the "attach to window" or "one aerial
window globally" approaches, the previous layout position of "group"
(pick the edge of the group of adjacent windows containing the source
buffer) is kind of an orphan. I have deprecated the config option,
though it will continue to work for now. I suspect very few people are
using it, so if no one complains then I will follow up by deleting the
old logic entirely.