fix!: ignore win/buffer behavior prevents opening aerial at all (#204)
Previously the ignore behavior was to not trigger open_automatic, and to
not update the aerial window when attach_mode = "global". Unfortunately,
that means that we end up in a weird state and it's difficult to make
the open/close/toggle behavior in ignored buffers consistent _unless_ we
prevent opening aerial altogether. This essentially makes ignored
buffers the same as unsupported, but with the key difference that it
won't trigger any of the close_automatic_events.

As part of this, we're also changing the default to NOT ignore unlisted
buffers. The main reason for this is that help buffers are unlisted by
default, but aerial supports them.