feat: experimental support for navigating to symbol names (#279)
Using the treesitter backend, we can use the same "selectionRange" logic
that LSP symbol sources use to provide more detailed information about
where the name of the symbol is. We already use the LSP information to
change how we navigate the cursor to symbols, so once we parse this
information from treesitter it will automatically get used.

I'm putting this behind an experimental config option for now so we can
test it out for a while before making a sudden change to the behavior.