fix(make): improve treesitter symbols query (#479)
Previously the entire list of targets was captured as a single symbol: instead capture each target.
vim's Makefile is an example that makes use of this feature.

Exclude targets that aren't targets from the "special target list": these are not symbols. This is a
minor improvement for Makefiles such as the aerial.nvim Makefile, which only declares `.PHONY` a
single time, but it is a large noise-reducer for cases such as nvim-treesitter Makefile, which lists
`.PHONY` before each phony target.

Signed-off-by: Robert Muir <rmuir@apache.org>