feat: treesitter queries for editorconfig, hcl, ssh_config (#484)
Terraform is a different treesitter parser than hcl, but for aerial purposes these are the same.
Support hcl also, it is helpful for docker bake files at least.

ssh_config / editorconfig are a few more "devops" formats with simple "sections", aerial helps when
the files are large. The editorconfig pattern is a bit strange, but this is due to the grammar:
any fancy stuff such as `not-kind-eq` is avoided since this predicate isn't in neovim core.

Signed-off-by: Robert Muir <rmuir@apache.org>